StartUpCommands: Port it from Yoinx

+++ app/src/main/java/com/grarak/kerneladiutor/utils/root/Control.java
@@ -40,23 +40,34 @@ public class Control implements Constants {
 
     public static void commandSaver(final Context context, final String path, final String command) {
         CommandDB commandDB = new CommandDB(context);
-
-        List<CommandDB.CommandItem> commandItems = commandDB.getAllCommands();
-        for (int i = 0; i < commandItems.size(); i++) {
-            String p = commandItems.get(i).getPath();
-            if (p != null && p.equals(path))
-                commandDB.delete(i);
+        // Something keeps trying to save commands wtih a null path... This makes the bootservice Force Close.
+        // Ensure that this isn't a possibility by not saving null paths or commands.
+        if (path != null && command != null && !path.equals("null") && !command.equals("null")) {
+            List<CommandDB.CommandItem> commandItems = commandDB.getAllCommands();
+            for (int i = 0; i < commandItems.size(); i++) {
+                String p = commandItems.get(i).getPath();
+                if (p != null && p.equals(path)) {
+                    commandDB.delete(i);
+                }
+            }
+            commandDB.putCommand(path, command);
+            commandDB.commit();
+        } else {
+            Log.i(TAG, "Unable to save command due to null values.");
         }
-
-        commandDB.putCommand(path, command);
-        commandDB.commit();
     }
 
     private static void run(String command, String path, Context context) {
-        RootUtils.runCommand(command);
-        commandSaver(context, path, command);
-        Log.i(TAG, "Run command: " + command);
+        if (path != null && command != null && !path.equals("null") && !command.equals("null")) {
+            RootUtils.runCommand(command);
+            commandSaver(context, path, command);
+            Log.i(TAG, "Run command: " + command);
+        }
+        else {
+            Log.i(TAG, "Unable to run command due to null values.");
+        }
     }
+
 
     private static int getChecksum(int arg1, int arg2) {
         return 255 & (Integer.MAX_VALUE ^ (arg1 & 255) + (arg2 & 255));
